docker demo improvements and settings reorganisation
