+++ .gitignore
@@ -1,5 +1,4 @@
 *.calltree
-*.conf
 *.egg-info/
 *.log
 *.prof
@@ -45,3 +44,4 @@ tardis/tardis_portal/filters/*
 !tardis/tardis_portal/filters/__init__.py
 !tardis/tardis_portal/filters/diffractionimage.py
 !tardis/tardis_portal/filters/jeolsem.py
+wheelhouse
