+++ docker-demo.env
@@ -1,0 +1,28 @@
+# Environment variables used in docker-compose.yml
+POSTGRES_PASSWORD=postgres
+
+DJANGO_SETTINGS_MODULE=tardis.env_settings
+
+# SECURITY WARNING: Don't run with DEBUG turned on in production!
+MYTARDIS_DEBUG=False
+
+# SECURITY WARNING: Keep the secret key used in production secret!
+MYTARDIS_DJANGO_SECRET_KEY=am#e4*nbH+4n@rt%ux6se1dV3_%f%8jpyua^wsh8-@rx9g6ryl
+
+# SECURITY WARNING: Keep the rabbitmq password used in production secret!
+MYTARDIS_BROKER_URL=amqp://guest:guest@rabbitmq:5672//
+
+MYTARDIS_SITE_TITLE=Dockered MyTardis
+MYTARDIS_SPONSORED_TEXT=Docker - Build, Ship, and Run Any App, Anywhere
+MYTARDIS_DEFAULT_INSTITUTION=MyTardis University
+
+# SECURITY WARNING: Keep the MyTardis admin password used in production secret!
+MYTARDIS_ADMIN_USERNAME=demo
+MYTARDIS_ADMIN_EMAIL=demo@example.com
+MYTARDIS_ADMIN_PASSWORD=demo
+
+# Gunicorn number of workers
+WEB_CONCURRENCY=2
+
+# Elasticsearch
+MYTARDIS_ES_HOST=elasticsearch
