--- docker-mysql.env
@@ -1,25 +1,0 @@
-# Environment variables used in docker-compose.yml for MySQL
-
-DJANGO_SETTINGS_MODULE=tardis.docker_settings
-
-# SECURITY WARNING: Don't run with DEBUG turned on in production!
-DEBUG=True
-
-# SECURITY WARNING: Keep the secret key used in production secret!
-DJANGO_SECRET_KEY=am#e4*nbH+4n@rt%ux6se1dV3_%f%8jpyua^wsh8-@rx9g6ryl
-
-DATABASE_URL=mysql://root:mysql@mysql/mysql
-# SECURITY WARNING: Keep the database password used in production secret!
-MYSQL_ROOT_PASSWORD=mysql
-
-# SECURITY WARNING: Keep the rabbitmq password used in production secret!
-BROKER_URL=amqp://guest:guest@rabbitmq:5672//
-
-SITE_TITLE=Dockered MyTardis
-SPONSORED_TEXT=Docker - Build, Ship, and Run Any App, Anywhere
-DEFAULT_INSTITUTION=MyTardis University
-
-# SECURITY WARNING: Keep the MyTardis admin password used in production secret!
-MYTARDIS_ADMIN_USERNAME=demo
-MYTARDIS_ADMIN_EMAIL=demo@example.com
-MYTARDIS_ADMIN_PASSWORD=demo
