+++ tardis/default_settings/admins.py
@@ -1,0 +1,2 @@
+ADMINS = (('bob', 'bob@bobmail.com'),)
+MANAGERS = ADMINS
