+++ tardis/default_settings/debug.py
@@ -1,0 +1,8 @@
+DEBUG = True
+"""Set to false for production use"""
+
+ALLOWED_HOSTS = ['*']
+'''
+For security reasons this needs to be set to your hostname and/or IP
+address in production.
+'''
