+++ tardis/default_settings/email.py
@@ -1,0 +1,11 @@
+# Email Configuration
+
+EMAIL_PORT = 587
+
+EMAIL_HOST = 'smtp.gmail.com'
+
+EMAIL_HOST_USER = 'bob@bobmail.com'
+
+EMAIL_HOST_PASSWORD = 'bob'
+
+EMAIL_USE_TLS = True
