+++ tardis/default_settings/i18n.py
@@ -1,0 +1,3 @@
+# If you set this to False, Django will make some optimizations so as not
+# to load the internationalization machinery.
+USE_I18N = True
