+++ tardis/default_settings/urls.py
@@ -1,0 +1,9 @@
+ROOT_URLCONF = 'tardis.urls'
+
+# URL that handles the media served from MEDIA_ROOT. Make sure to use a
+# trailing slash if there is a path component (optional in other cases).
+# Examples: "http://media.lawrence.com", "http://example.com/media/"
+MEDIA_URL = None
+
+# Static content location
+STATIC_URL = '/static/'
